:sparkles: feat: add context mixin support
